feat(types): centralize go domain types
