+++ .github/issue-updates/15db0fe8-653f-4eae-aa25-915296f4728a.json
@@ -1,0 +1,7 @@
+{
+  "action": "comment",
+  "number": 1,
+  "body": "Working on central types in pkg/types",
+  "guid": "15db0fe8-653f-4eae-aa25-915296f4728a",
+  "legacy_guid": "comment-issue-1-2025-07-05"
+}
